+++ src/ufo/obslocalization/ObsVertLocalization.h
@@ -217,7 +217,7 @@ ObsVertLocalization<MODEL>::getLocalObs(const GeometryIterator_ & i,
   }
   // truncate to maxnobs if needed
   const boost::optional<int> & maxnobs = options_.maxnobs;
-  if ( (maxnobs != boost::none) && (localobs.index.size() > *maxnobs ) ) {
+  if ( (maxnobs != boost::none) && (static_cast<int>(localobs.index.size()) > *maxnobs ) ) {
     for (unsigned int jj = 0; jj < localobs.index.size(); ++jj) {
         oops::Log::debug() << "Before sort [i, d]: " << localobs.index[jj]
             << " , " << localobs.distance[jj] << std::endl;
